llm/client_server_openai/backend/pyproject.toml
[tool.poetry]
name = "openai_backend"
version = "0.1.0"
description = ""
authors = ["Vicente Herrera <vicenteherrera@vicenteherrera.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

# [[tool.poetry.source]]
# name = "devpi"
# url = "http://127.0.0.1:3141/root/pypi/+simple/"
# priority = "primary"
# Disable PyPI and uncomment this to use devpi local library cache

[tool.poetry.dependencies]
python = "3.11.7"
# openai = "^1.30.1"
python-dotenv = "^1.0.1"
jsonify = "^0.5"
flask = "^3.0.3"
langchain-openai = "^0.1.8"
# langchain-community = "^0.2.4"
# langchain = "^0.2.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
mypy = "^1.7.0"
black = "^23.11.0"

[build-system]
requires = ["poetry-core>=1.8.3"]
build-backend = "poetry.core.masonry.api"
